base components
the minimum navigation/routing is implemented, along with the creation of components for the main (not gated by validation) pages and reusable components (navbars and footer). Creating forms to integrate with backend is the next step. Step after that is implementing dynamic rendering to show either signup/login buttons or loggedinnavbar depending on authentication. CSS can come later.