Reviews and setting up for replies
Reviews are mostly finalized here, but there are a couple things left to add, such as options to edit and delete if the logged in user is the author of the review.

I also put together a preliminary model for replies, which is what the objects representing responses to reviews and other replies will be called. I would have named this "Comment" instead, but comment was already taken by the name of the key in the review schema, and this makes a similar amount of sense.