Reviews
Reviews are functioning properly, but there is more to build out, including editing and deleting routes and UI. Timestamps should also be added to be able to display reviews from newest to oldest, displaying the date it was posted.

Next is building out the replies, which will be similar to the reviews, but with some exclusions such as the rating.